Move the viz controls out
