+++ packages/transform-dataresource/src/charts/grid.js
@@ -1,0 +1,20 @@
+import * as React from "react";
+import VirtualizedGrid from "../virtualized-grid";
+
+export const DataResourceTransformGrid = ({
+  data: { data, schema },
+  theme,
+  expanded,
+  height
+}) => {
+  return (
+    <VirtualizedGrid
+      data={data}
+      schema={schema}
+      theme={theme}
+      expanded={expanded}
+      height={height}
+      // style={{ marginRight: "10px" }}
+    />
+  );
+};
