+++ packages/transform-dataresource/src/index.js
@@ -2,11 +2,12 @@
 import { hot } from "react-hot-loader";
 
 import * as React from "react";
-import VirtualizedGrid from "./virtualized-grid";
 import { DatabaseOcticon, Beaker } from "@nteract/octicons";
 
 import { colors } from "./settings";
-import { semioticSettings } from "./charts";
+import { semioticSettings } from "./charts/settings";
+import { DataResourceTransformGrid } from "./charts/grid";
+import VizControls from "./VizControls";
 
 import {
   TreeIcon,
@@ -78,67 +79,6 @@ const generateChartKey = ({
   )}-${pieceType}-${summaryType}-${networkType}-${hierarchyType}-${JSON.stringify(
     chart
   )}`;
-
-const DataResourceTransformGrid = ({
-  data: { data, schema },
-  theme,
-  expanded,
-  height
-}) => {
-  return (
-    <VirtualizedGrid
-      data={data}
-      schema={schema}
-      theme={theme}
-      expanded={expanded}
-      height={height}
-      // style={{ marginRight: "10px" }}
-    />
-  );
-};
-
-const availableLineTypes = [
-  {
-    type: "line",
-    label: "Line Chart"
-  },
-  {
-    type: "stackedarea",
-    label: "Stacked Area Chart"
-  },
-  {
-    type: "stackedpercent",
-    label: "Stacked Area Chart (Percent)"
-  },
-  {
-    type: "bumparea",
-    label: "Ranked Area Chart"
-  }
-];
-
-const metricDimSelector = (
-  values,
-  selectionFunction,
-  title,
-  required,
-  selectedValue
-) => {
-  return (
-    <div style={{ display: "inline-block", margin: "0 10px" }}>
-      <h2>{title}</h2>
-      <select
-        value={selectedValue}
-        onChange={e => selectionFunction(e.target.value)}
-      >
-        {(required ? values : ["none", ...values]).map(d => (
-          <option key={`selector-option-${d}`} value={d} label={d}>
-            {d}
-          </option>
-        ))}
-      </select>
-    </div>
-  );
-};
 
 /*
   contour is an option for scatterplot
@@ -341,178 +281,19 @@ class DataResourceTransform extends React.Component<Props, State> {
           size={[500, height - 200]}
           {...frameSettings}
         />
-        {(view === "summary" ||
-          view === "scatter" ||
-          view === "hexbin" ||
-          view === "bar" ||
-          view === "network" ||
-          view === "hierarchy") &&
-          metricDimSelector(
-            metrics.map(d => d.name),
-            d => this.updateChart({ chart: { ...chart, metric1: d } }),
-            view === "scatter" || view === "hexbin" ? "X" : "Metric",
-            true,
-            chart.metric1
-          )}
-        {(view === "scatter" || view === "hexbin") &&
-          metricDimSelector(
-            metrics.map(d => d.name),
-            d => this.updateChart({ chart: { ...chart, metric2: d } }),
-            "Y",
-            true,
-            chart.metric2
-          )}
-        {(view === "scatter" || view === "bar") &&
-          metricDimSelector(
-            metrics.map(d => d.name),
-            d => this.updateChart({ chart: { ...chart, metric3: d } }),
-            view === "bar" ? "WIDTH" : "SIZE",
-            false,
-            chart.metric3
-          )}
-        {(view === "summary" ||
-          view === "scatter" ||
-          view === "bar" ||
-          view === "parallel") &&
-          metricDimSelector(
-            dimensions.map(d => d.name),
-            d => this.updateChart({ chart: { ...chart, dim1: d } }),
-            view === "summary" ? "CATEGORY" : "COLOR",
-            true,
-            chart.dim1
-          )}
-        {view === "scatter" &&
-          metricDimSelector(
-            dimensions.map(d => d.name),
-            d => this.updateChart({ chart: { ...chart, dim2: d } }),
-            "LABELS",
-            false,
-            chart.dim2
-          )}
-        {view === "network" &&
-          metricDimSelector(
-            dimensions.map(d => d.name),
-            d => this.updateChart({ chart: { ...chart, dim1: d } }),
-            "SOURCE",
-            true,
-            chart.dim1
-          )}
-        {view === "network" &&
-          metricDimSelector(
-            dimensions.map(d => d.name),
-            d => this.updateChart({ chart: { ...chart, dim2: d } }),
-            "TARGET",
-            true,
-            chart.dim2
-          )}
-        {view === "network" &&
-          metricDimSelector(
-            ["force", "sankey"],
-            d => this.updateChart({ networkType: d }),
-            "TYPE",
-            true,
-            networkType
-          )}
-        {view === "hierarchy" &&
-          metricDimSelector(
-            ["dendrogram", "treemap", "partition"],
-            d => this.updateChart({ hierarchyType: d }),
-            "TYPE",
-            true,
-            hierarchyType
-          )}
-        {view === "summary" &&
-          metricDimSelector(
-            ["violin", "boxplot", "joy", "heatmap", "histogram"],
-            d => this.updateChart({ summaryType: d }),
-            "TYPE",
-            true,
-            summaryType
-          )}
-        {view === "line" &&
-          metricDimSelector(
-            ["array-order", ...metrics.map(d => d.name)],
-            d => this.updateChart({ chart: { ...chart, timeseriesSort: d } }),
-            "Sort by",
-            true,
-            chart.timeseriesSort
-          )}
-        {view === "line" && (
-          <div style={{ display: "inline-block" }}>
-            <h2>Chart Type</h2>
-            {availableLineTypes.map(d => (
-              <button
-                style={{
-                  marginLeft: "0px",
-                  color: lineType === d.type ? "lightgray" : "black"
-                }}
-                onClick={() => this.setLineType(d.type)}
-              >
-                {d.label}
-              </button>
-            ))}
-          </div>
-        )}
-        {view === "hierarchy" && (
-          <div
-            style={{
-              display: "inline-block",
-              width: "30%",
-              marginLeft: "20px"
-            }}
-          >
-            <h2>Nesting</h2>
-            {selectedDimensions.length === 0
-              ? "Select categories to nest"
-              : `root, ${selectedDimensions.join(", ")}`}
-          </div>
-        )}
-        {(view === "bar" || view === "hierarchy") && (
-          <div
-            style={{
-              display: "inline-block",
-              width: "30%",
-              marginLeft: "20px"
-            }}
-          >
-            <h2>Categories</h2>
-            {dimensions.map(d => (
-              <button
-                key={`dimensions-select-${d.name}`}
-                style={{
-                  marginLeft: "20px",
-                  color:
-                    selectedDimensions.indexOf(d.name) !== -1
-                      ? "black"
-                      : "lightgray"
-                }}
-                onClick={() => this.updateDimensions(d.name)}
-              >
-                {d.name}
-              </button>
-            ))}
-          </div>
-        )}
-        {view === "line" && (
-          <div style={{ display: "inline-block" }}>
-            <h2>Metrics</h2>
-            {metrics.map(d => (
-              <button
-                key={`metrics-select-${d.name}`}
-                style={{
-                  marginLeft: "0px",
-                  color:
-                    selectedMetrics.indexOf(d.name) !== -1
-                      ? "black"
-                      : "lightgray"
-                }}
-                onClick={() => this.updateMetrics(d.name)}
-              >
-                {d.name}
-              </button>
-            ))}
-          </div>
-        )}
+        <VizControls
+          {...{
+            view,
+            chart,
+            metrics,
+            dimensions,
+            selectedDimensions,
+            hierarchyType,
+            summaryType,
+            networkType,
+            updateChart: this.updateChart
+          }}
+        />
         <style jsx>{`
           :global(.tooltip-content) {
             color: black;
@@ -639,7 +420,6 @@ class DataResourceTransform extends React.Component<Props, State> {
     const {
       view,
       dimensions,
-      metrics,
       chart,
       lineType,
       selectedDimensions,
@@ -647,12 +427,8 @@ class DataResourceTransform extends React.Component<Props, State> {
       pieceType,
       summaryType,
       networkType,
-      hierarchyType,
-      primaryKey,
-      data: stateData
+      hierarchyType
     } = this.state;
-
-    const { data, height } = this.props;
 
     let display = null;
 
@@ -670,8 +446,6 @@ class DataResourceTransform extends React.Component<Props, State> {
         "parallel"
       ].includes(view)
     ) {
-      const { Frame, chartGenerator } = semioticSettings[view];
-
       const chartKey = generateChartKey({
         view,
         lineType,
